requirements.txt
anyascii==0.1.7
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.5
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
l18n==2020.6.1
openpyxl==3.0.6
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
Pygments==2.7.3
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
traitlets==5.0.5
Unidecode==1.1.2
urllib3==1.26.2
wagtail==2.11.3
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
